refactor (Day4): Extract each rule checker out into their own file
- Remove exports from index.js
 - No need to export the functions anymore since they have their own files now
